Fix: restrict program refs to item hashes
Problem: some users specify nonsensical strings in the ref field
of programs and program volumes.

Solution: enforce the use of item hashes with the `ItemHash` type.

Moved the `ItemHash` and `ItemType` definitions to their own module
to avoid circular imports.
